Migrate chart rendering out of @chart-io/core into React components/hooks
Inlines the per-chart-type render/focus/tooltip logic that used to live in
packages/core/src/renderers (bar, column, line, area, scatter) directly into
the corresponding Base components and hooks in packages/react. Removes the
now-empty renderers directory along with the IRenderProps type and
getParentKey util that only existed to support it, in favour of the
render logic living alongside the component that uses it.
