+++ packages/core/src/index.ts
@@ -2,7 +2,6 @@ export * from "./canvas";
 export * from "./constants";
 export * from "./d3";
 export * as detection from "./detection";
-export * from "./renderers";
 export * from "./store";
 export * from "./themes";
 export * from "./types";
