--- packages/core/src/renderers/column/index.ts
@@ -1,14 +1,0 @@
-import { grouped as groupedRender } from "./grouped";
-import { stacked as stackedRender } from "./stacked";
-
-export const stacked = {
-    render: stackedRender,
-};
-
-export const grouped = {
-    render: groupedRender,
-};
-
-export * from "./focus";
-export * from "./render";
-export * from "./tooltip";
