--- packages/core/src/renderers/index.ts
@@ -1,7 +1,0 @@
-import * as area from "./area";
-import * as bar from "./bar";
-import * as column from "./column";
-import * as line from "./line";
-import * as scatter from "./scatter";
-
-export { scatter, column, bar, line, area };
