--- packages/core/src/renderers/line/index.ts
@@ -1,9 +1,0 @@
-import { canvas as canvasRender } from "./canvas";
-
-export const canvas = {
-    render: canvasRender,
-};
-
-export * from "./render";
-export * from "./focus";
-export * from "./tooltip";
