--- packages/core/src/renderers/scatter/index.ts
@@ -1,3 +1,0 @@
-export * from "./focus";
-export * from "./render";
-export * from "./tooltip";
