+++ packages/react/src/lib/components/Plots/Area/StackedArea/StackedAreaBase.tsx
@@ -1,5 +1,4 @@
-import { area, chartSelectors, IState } from "@chart-io/core";
-import { d3 } from "@chart-io/core";
+import { chartSelectors, d3, ensureNoScaleOverflow, interpolateMultiPath, IState } from "@chart-io/core";
 import type { IColor, IEventPlotProps } from "@chart-io/core";
 
 import { useMemo } from "react";
@@ -7,6 +6,7 @@ import { useSelector } from "react-redux";
 
 import { useLegendItems, useRender } from "../../../../hooks";
 
+import { IBandwidthScale } from "../../IBandwidthScale";
 import { useDatumFocus } from "./useDatumFocus";
 import { useMultiPathCreator } from "./useMultiPathCreator";
 import { useTooltip } from "./useTooltip";
@@ -53,15 +53,78 @@ export function StackedAreaBase({
     useLegendItems(ys, "line", showInLegend, colors);
 
     /* On future renders we want to update the path */
-    useRender(async () => {
-        const props = { x, ys, xScale, yScale, data: sortedData, colors, theme };
+    useRender(() => {
+        const bandwidth = (xScale as IBandwidthScale).bandwidth ? (xScale as IBandwidthScale).bandwidth() / 2 : 0;
+
+        const areaShape = d3
+            .area()
+            .curve(d3.curveLinear)
+            // @ts-ignore: TODO: Not sure how to fix this
+            .x((d) => xScale(d.data[x]) + bandwidth)
+            .y0((d) => yScale(d[0]))
+            .y1((d) => yScale(d[1]));
 
         // Handle Canvas rendering
         if (canvas) {
-            return area.stacked.canvas.render({ ...props, width, height, canvas });
+            ensureNoScaleOverflow(yScale, sortedData, ys, "StackedSVGArea");
+
+            // @ts-ignore: TODO: Not sure how to fix this
+            const stackedData = d3.stack().keys(ys)(sortedData);
+            const colorScale = d3.scaleOrdinal().domain(ys).range(colors);
+
+            const context = canvas.getContext("2d");
+            areaShape.context(context);
+
+            // Create a join with a faux dom element to work out the areas we care about
+            const join = d3.select(document.createElement("custom")).selectAll("path").data(stackedData);
+
+            context.clearRect(0, 0, width, height);
+
+            join.enter().each((d) => {
+                const color = colorScale(d.key);
+                const fillColor = d3.color(color.toString());
+                fillColor.opacity = theme.series.opacity;
+                const strokeColor = fillColor.darker();
+
+                context.beginPath();
+
+                // @ts-ignore: TODO: Not sure how to fix this
+                areaShape(d);
+
+                context.fillStyle = fillColor.toString();
+                context.strokeStyle = strokeColor.toString();
+                context.fill();
+                context.stroke();
+            });
+
+            return;
         }
 
-        area.stacked.render({ ...props, layer: layer.current, animationDuration });
+        if (!layer.current) {
+            return;
+        }
+
+        ensureNoScaleOverflow(yScale, sortedData, ys, "StackedSVGArea");
+
+        // @ts-ignore: TODO: Not sure how to fix this
+        const stackedData = d3.stack().keys(ys)(sortedData);
+        const colorScale = d3.scaleOrdinal().domain(ys).range(colors);
+
+        const join = d3.select(layer.current).selectAll("path").data(stackedData);
+
+        join.style("fill", (d) => colorScale(d.key).toString())
+            .style("stroke", (d) => d3.color(colorScale(d.key).toString()).darker().toString())
+            .style("opacity", theme.series.opacity)
+            .style("pointer-events", "none")
+            .transition("area")
+            .duration(animationDuration)
+            .ease(d3.easeCubicInOut)
+            .attrTween("d", function (d) {
+                const previous = d3.select(this).attr("d");
+                // @ts-ignore: TODO: Not sure how to fix this
+                const current = areaShape(d);
+                return interpolateMultiPath(previous, current);
+            });
     }, [x, ys, sortedData, xScale, yScale, layer, animationDuration, theme.series.opacity]);
 
     // If possible respond to global mouse events for tooltips etc
