+++ packages/react/src/lib/components/Plots/Area/StackedArea/useDatumFocus.ts
@@ -1,13 +1,134 @@
-import { area, eventSelectors, IState } from "@chart-io/core";
+import { d3, eventActions, eventSelectors, getDistance, IState } from "@chart-io/core";
+import type { IColor, ICoordinate, IData, IDispatch, IInvertScale, IMouseEventType, INumericValue, IScale } from "@chart-io/core";
 
 import { useSelector, useStore } from "react-redux";
 import { useEffect } from "react";
 
-interface IDatumsFocusProps extends Omit<area.IStackedAreaFocusProps, "dispatch" | "eventMode" | "position"> {
+interface IDatumsFocusProps {
     /**
      * Whether the plot is interactive
      */
     interactive: boolean;
+    /**
+     * The key of the field used for the x position
+     */
+    x: string;
+    /**
+     * The keys of the fields used for the y positions
+     */
+    ys: string[];
+    /**
+     * The d3 scale for the x axis
+     */
+    xScale: IScale | undefined;
+    /**
+     * The d3 scale for the y axis
+     */
+    yScale: IScale | undefined;
+    /**
+     * The complete dataset for the plot
+     */
+    data: IData;
+    /**
+     * The color of the plot
+     */
+    colors: IColor[];
+}
+
+interface IStackedFocusParams {
+    dispatch: IDispatch;
+    x: string;
+    ys: string[];
+    colors: IColor[];
+    xScale: IScale | undefined;
+    yScale: IScale | undefined;
+    position: ICoordinate;
+    data: IData;
+    eventMode: IMouseEventType;
+}
+
+const defaultValues = {
+    markers: [],
+    horizontalDroplines: [],
+    verticalDroplines: [],
+    sum: 0,
+};
+
+/**
+ * Helper function to manage markers & droplines for a selected datum on the Area plot
+ * @return              A function to set the focused datum
+ */
+function focus({ dispatch, x, ys, xScale, yScale, position, data, colors, eventMode }: IStackedFocusParams) {
+    if (!xScale || !yScale || !data) {
+        // istanbul ignore next
+        return;
+    }
+
+    // We need the invert function to be able to do this
+    if (!(xScale as IInvertScale).invert) {
+        // istanbul ignore next
+        return;
+    }
+
+    // Don't run for the EXIT event
+    if (eventMode === "NONE") {
+        return;
+    }
+
+    // Work out the datum that we're closet to
+    const xValue = (xScale as IInvertScale).invert(position.x);
+    const index = d3.bisector((d) => d[x]).center(data, xValue);
+    const datum = data[index];
+
+    // Get the appropriate attributes
+    const cx = xScale(datum[x] as INumericValue);
+
+    const { markers, horizontalDroplines, verticalDroplines } = ys.reduce((result, y, i) => {
+        // @ts-ignore TODO: Not sure how to fix this
+        const sum = result.sum + datum[y];
+        const cy = yScale(sum);
+        const color = d3.color(colors[i].toString()).darker();
+
+        const distance = getDistance(position.x, position.y, cx, cy);
+
+        const marker = { fill: color, r1: 5, r2: 5, cx, cy, distance };
+        const horizontalDropline = {
+            isHorizontal: true,
+            color,
+            x1: cx,
+            x2: xScale.range()[0],
+            y1: cy,
+            y2: cy,
+            distance,
+        };
+        const verticalDropline = {
+            isVertical: true,
+            color,
+            x1: cx,
+            x2: cx,
+            y1: cy,
+            y2: yScale.range()[0],
+            distance,
+        };
+        return {
+            sum,
+            markers: [...result.markers, marker],
+            horizontalDroplines: [...result.horizontalDroplines, horizontalDropline],
+            verticalDroplines: [...result.verticalDroplines, verticalDropline],
+        };
+    }, defaultValues);
+
+    markers.forEach((marker) => dispatch(eventActions.addMarker(marker)));
+    horizontalDroplines.forEach((horizontalDropline) => dispatch(eventActions.addDropline(horizontalDropline)));
+    verticalDroplines.forEach((verticalDropline) => dispatch(eventActions.addDropline(verticalDropline)));
+
+    // Clean up operations on exit
+    // prettier-ignore
+    return () => {
+        markers.forEach((marker) => dispatch(eventActions.removeMarker(marker)));
+        horizontalDroplines.forEach((horizontalDropline) => dispatch(eventActions.removeDropline(horizontalDropline)));
+        verticalDroplines.forEach((verticalDropline) => dispatch(eventActions.removeDropline(verticalDropline)));
+    };
 }
 
 /**
@@ -22,7 +143,7 @@ export function useDatumFocus({ interactive, x, ys, xScale, yScale, data, colors
     useEffect(() => {
         /* If possible respond to global mouse events for tooltips etc */
         if (interactive) {
-            return area.stacked.focus({
+            return focus({
                 dispatch,
                 x,
                 ys,
