+++ packages/react/src/lib/components/Plots/Bar/StackedBar/StackedBarBase.tsx
@@ -1,5 +1,13 @@
-import { bar, chartSelectors, d3, IState } from "@chart-io/core";
-import type { IColor, IEventPlotProps } from "@chart-io/core";
+import {
+    chartSelectors,
+    d3,
+    ensureBandwidth,
+    ensureNoScaleOverflow,
+    ensureValuesAreUnique,
+    getBandwidthAndOffset,
+    IState,
+} from "@chart-io/core";
+import type { IColor, IDatum, IEventPlotProps } from "@chart-io/core";
 
 import { useSelector } from "react-redux";
 
@@ -8,6 +16,7 @@ import { useLegendItems, useRender } from "../../../../hooks";
 import { renderCanvas } from "../../renderCanvas";
 import { useFocused } from "../useFocused";
 import { useTooltip } from "../useTooltip";
+import { getParentKey } from "./getParentKey";
 
 export interface IStackedBarBaseProps extends Omit<IEventPlotProps, "ys" | "x"> {
     /**
@@ -62,23 +71,85 @@ export function StackedBarBase({
 
     // prettier-ignore
     useRender(() => {
-        const { update } = bar.stacked.render({
-            animationDuration,
-            interactive,
-            layer: layer.current,
-            data,
-            colors,
-            onMouseOver,
-            onMouseOut,
-            onClick,
-            onFocus,
-            onTooltip,
-            theme,
-            xs,
-            y,
-            xScale,
-            yScale,
-        });
+        const { bandwidth, offset } = getBandwidthAndOffset(yScale, y, data);
+
+        if (ensureBandwidth(bandwidth, "StackedBar") === false) return;
+        ensureValuesAreUnique(data, y, "StackedBar");
+        ensureNoScaleOverflow(xScale, data, xs, "StackedBar");
+
+        // Create the stacked variant of the data
+        const keys = xs;
+        // @ts-ignore: TODO: Work out how to fix this
+        const stackedData = d3.stack().keys(keys)(data);
+        const colorScale = d3.scaleOrdinal().domain(keys).range(colors);
+
+        const groupJoin = d3.select(layer.current).selectAll<SVGGElement, IDatum>("g").data(stackedData);
+
+        // Clean up old stacks
+        groupJoin.exit().remove();
+
+        const join = groupJoin
+            .enter()
+            .append("g")
+            // @ts-ignore: TODO: Work out how to fix this
+            .merge(groupJoin)
+            .selectAll(".bar")
+            .data((d) => d);
+
+        join.exit().remove();
+        const enter = join
+            .enter()
+            .append("rect")
+            .attr("class", "bar")
+            .attr("x", () => xScale.range()[0] as number)
+            .attr("y", (d) => yScale(d.data[y]) - offset)
+            .attr("height", bandwidth)
+            .attr("width", 0)
+            .style("fill", (d, i, elements) => {
+                const key = getParentKey(elements[i]);
+                return colorScale(key)?.toString();
+            })
+            .style("opacity", theme.series.opacity);
+
+        const update = join
+            .merge(enter)
+            .style("opacity", theme.series.opacity)
+            .on("mouseover", function (event: MouseEvent, d) {
+                // istanbul ignore next
+                if (!interactive) return;
+
+                onMouseOver && onMouseOver(d.data, this as Element, event);
+                onFocus && onFocus({ element: this as Element, event, datum: d.data });
+                onTooltip && onTooltip({ datum: d.data, event, fillColors: xs.map((x) => colorScale(x) as IColor), xs });
+            })
+            .on("mouseout", function (event: MouseEvent, d) {
+                // istanbul ignore next
+                if (!interactive) return;
+
+                onMouseOut && onMouseOut(d.data, this as Element, event);
+                onFocus && onFocus(null);
+                onTooltip && onTooltip(null);
+            })
+            .on("click", function (event: MouseEvent, d: { data: IDatum }) {
+                // istanbul ignore next
+                if (!interactive) return;
+
+                onClick && onClick(d.data, this as Element, event);
+            })
+            .transition("position")
+            .duration(animationDuration / 2)
+            .style("fill", (d, i, elements) => {
+                const key = getParentKey(elements[i] as Element);
+                return colorScale(key)?.toString();
+            })
+            .attr("y", (d) => yScale(d.data[y]) - offset)
+            .attr("height", bandwidth)
+            // @ts-expect-error: Looks like the type defs are wrong missing named transitions
+            .transition("width")
+            .duration(animationDuration / 2)
+            .delay(animationDuration / 2)
+            .attr("width", (d) => xScale(d[1]) - xScale(d[0]))
+            .attr("x", (d) => xScale(d[0]));
 
         // @ts-ignore: TODO: Fix this TS
         renderCanvas(canvas, renderVirtualCanvas, width, height, update);
