+++ packages/react/src/lib/components/Plots/Column/Column/ColumnBase.tsx
@@ -1,5 +1,5 @@
-import { chartSelectors, column, d3, IState } from "@chart-io/core";
-import { IEventPlotProps } from "@chart-io/core";
+import { chartSelectors, d3, ensureBandwidth, ensureValuesAreUnique, getBandwidthAndOffset, IState } from "@chart-io/core";
+import type { IEventPlotProps, INumericValue } from "@chart-io/core";
 
 import { useSelector } from "react-redux";
 
@@ -51,23 +51,70 @@ export function ColumnBase({
     const onFocus = useFocused({ xScale, theme, grouped: false });
 
     useRender(() => {
-        const { update } = column.render({
-            animationDuration,
-            interactive,
-            layer: layer.current,
-            data,
-            fillColor,
-            onMouseOver,
-            onMouseOut,
-            onClick,
-            onFocus,
-            onTooltip,
-            theme,
-            x,
-            y,
-            xScale,
-            yScale,
-        });
+        const { bandwidth, offset } = getBandwidthAndOffset(xScale, x, data);
+
+        if (ensureBandwidth(bandwidth, "Column") === false) return;
+        ensureValuesAreUnique(data, x, "Column");
+
+        // D3 data join
+        const join = d3
+            .select(layer.current)
+            .selectAll<SVGRectElement, any>(".column")
+            .data(data, (d) => d[x]?.toString());
+
+        // Exit bars
+        join.exit().remove();
+
+        // Enter bars
+        const enter = join
+            .enter()
+            .append("rect")
+            .attr("class", "column")
+            // @ts-ignore: TODO: Need to work out casting
+            .attr("x", (d) => xScale(d[x]) - offset)
+            .attr("y", yScale.range()[0] as number)
+            .attr("width", bandwidth)
+            .attr("height", 0)
+            .style("fill", fillColor.toString());
+
+        // Update new and existing points
+        const update = enter
+            .merge(join)
+            .style("opacity", theme.series.opacity)
+            .on("mouseover", function (event, datum) {
+                // istanbul ignore next
+                if (!interactive) return;
+
+                onMouseOver && onMouseOver(datum, this, event);
+                onFocus && onFocus({ element: this, event, datum });
+                onTooltip && onTooltip({ datum, event, fillColors: [fillColor], ys: [y] });
+            })
+            .on("mouseout", function (event, datum) {
+                // istanbul ignore next
+                if (!interactive) return;
+
+                onMouseOut && onMouseOut(datum, this, event);
+                onFocus && onFocus(null);
+                onTooltip && onTooltip(null);
+            })
+            .on("click", function (event, datum) {
+                // istanbul ignore next
+                if (!interactive) return;
+
+                onClick && onClick(datum, this, event);
+            })
+            .transition("position")
+            .duration(animationDuration / 2)
+            // @ts-ignore: TODO: Need to work out casting
+            .attr("x", (d) => xScale(d[x]) - offset)
+            .style("fill", fillColor.toString())
+            .attr("width", bandwidth)
+            // @ts-expect-error: Looks like the type defs are wrong missing named transitions
+            .transition("height")
+            .duration(animationDuration / 2)
+            .delay(animationDuration / 2)
+            .attr("y", (d) => yScale(d[y] as INumericValue))
+            .attr("height", (d) => (yScale.range()[0] as number) - yScale(d[y] as INumericValue));
 
         renderCanvas(canvas, renderVirtualCanvas, width, height, update);
     }, [
