+++ packages/react/src/lib/components/Plots/Column/GroupedColumn/GroupedColumnBase.tsx
@@ -1,5 +1,5 @@
-import { chartSelectors, column, d3, IState } from "@chart-io/core";
-import type { IColor, IEventPlotProps } from "@chart-io/core";
+import { chartSelectors, d3, ensureBandwidth, getBandwidthAndOffset, IState } from "@chart-io/core";
+import type { IColor, IDatum, IEventPlotProps, INumericValue } from "@chart-io/core";
 
 import { useSelector } from "react-redux";
 
@@ -58,23 +58,74 @@ export function GroupedColumnBase({
     useLegendItems(ys, "square", showInLegend, colors);
 
     useRender(() => {
-        const { update } = column.grouped.render({
-            animationDuration,
-            interactive,
-            layer: layer.current,
-            data,
-            colors,
-            onMouseOver,
-            onMouseOut,
-            onClick,
-            onFocus,
-            onTooltip,
-            theme,
-            x,
-            ys,
-            xScale,
-            yScale,
-        });
+        const { bandwidth, offset } = getBandwidthAndOffset(xScale, x, data);
+
+        if (ensureBandwidth(bandwidth, "GroupedColumn") === false) return;
+
+        // Create a scale for each series to fit along the x-axis and the series colors
+        const colorScale = d3.scaleOrdinal().domain(ys).range(colors);
+        const x1Scale = d3.scaleBand().domain(ys).rangeRound([0, bandwidth]).padding(0.05);
+
+        const groupJoin = d3.select(layer.current).selectAll<SVGGElement, IDatum>("g").data(data);
+
+        // Clean up old groups
+        groupJoin.exit().remove();
+
+        const join = groupJoin
+            .enter()
+            .append("g")
+            .merge(groupJoin)
+            .selectAll(".column")
+            .data((d) => ys.map((y) => ({ ...d, key: y, value: d[y] })));
+
+        join.exit().remove();
+        const enter = join
+            .enter()
+            .append("rect")
+            .attr("class", "column")
+            .attr("x", (d) => xScale(d[x]) + x1Scale(d.key) - offset)
+            .attr("y", yScale.range()[0] as number)
+            .attr("height", 0)
+            .attr("width", x1Scale.bandwidth())
+            .style("fill", (d) => colorScale(d.key).toString())
+            .style("opacity", theme.series.opacity);
+
+        const update = join
+            .merge(enter)
+            .style("opacity", theme.series.opacity)
+            .on("mouseover", function (event, datum) {
+                // istanbul ignore next
+                if (!interactive) return;
+
+                onMouseOver && onMouseOver(datum, this as Element, event);
+                onFocus && onFocus({ element: this as Element, event, datum });
+                onTooltip && onTooltip({ datum, event, fillColors: [colorScale(datum.key) as IColor], ys: [datum.key] });
+            })
+            .on("mouseout", function (event, datum) {
+                // istanbul ignore next
+                if (!interactive) return;
+
+                onMouseOut && onMouseOut(datum, this as Element, event);
+                onFocus && onFocus(null);
+                onTooltip && onTooltip(null);
+            })
+            .on("click", function (event, datum) {
+                // istanbul ignore next
+                if (!interactive) return;
+
+                onClick && onClick(datum, this as Element, event);
+            })
+            .transition("position")
+            .duration(animationDuration / 2)
+            .attr("x", (d) => xScale(d[x]) + x1Scale(d.key) - offset)
+            .attr("width", x1Scale.bandwidth())
+            .style("fill", (d) => colorScale(d.key).toString())
+            // @ts-expect-error: Looks like the type defs are wrong missing named transitions
+            .transition("height")
+            .duration(animationDuration / 2)
+            .delay(animationDuration / 2)
+            .attr("height", (d) => (yScale.range()[0] as number) - yScale(d.value as INumericValue))
+            .attr("y", (d) => yScale(d.value as INumericValue));
 
         // @ts-ignore: TODO: Work out how to fix this
         renderCanvas(canvas, renderVirtualCanvas, width, height, update);
