+++ packages/react/src/lib/components/Plots/Column/StackedColumn/getParentKey.ts
@@ -1,9 +1,9 @@
 /**
  * Returns the key of the parent element (used for Stacked series)
- * @param  {HTMLElement} element    The element
- * @param  {Number} i               The index of the element
- * @return {String}                 The key used by the parent
+ * @param  element    The element
+ * @return            The key used by the parent
  */
-export const getParentKey = (element) => {
+export function getParentKey(element: Element): string {
+    // @ts-ignore: TODO: How do we protect against this D3 internal stuff?
     return element?.parentNode?.__data__?.key;
 }
