+++ packages/react/src/lib/components/Plots/Column/useTooltip.ts
@@ -1,8 +1,35 @@
-import { IColor, IDatum, IMouseEvent } from "@chart-io/core";
-import { column } from "@chart-io/core";
+import { eventActions } from "@chart-io/core";
+import type { IColor, IDatum, IDispatch, IMouseEvent } from "@chart-io/core";
 
 import { useEffect, useState } from "react";
 import { useStore } from "react-redux";
+
+export interface IColumnTooltipProps {
+    /**
+     * The redux store dispatch function
+     */
+    dispatch: IDispatch;
+    /**
+     * The focused data point
+     */
+    datum: IDatum;
+    /**
+     * The color of the data point
+     */
+    colors: IColor[];
+    /**
+     * The event that initiated the request to display the tooltip
+     */
+    event: IMouseEvent;
+    /**
+     * The key of the field used for the x position
+     */
+    x: string;
+    /**
+     * The keys of the fields used for the y position
+     */
+    ys: string[];
+}
 
 interface ITooltipParams {
     datum: IDatum;
@@ -12,12 +39,56 @@ interface ITooltipParams {
 }
 
 /**
+ * Helper function to manage tooltips for a selected datum on the Column plot
+ * @return              A function to set the focused datum
+ */
+function tooltip({ dispatch, x, ys, datum, colors, event }: IColumnTooltipProps) {
+    if (!datum) return;
+
+    // Only use border colors for a single item
+    if (colors && colors.length === 1) {
+        dispatch(eventActions.setTooltipBorderColor(colors[0]));
+    }
+
+    // Common x value
+    const tooltipItemX = {
+        datum,
+        name: x,
+        value: datum[x],
+    };
+    dispatch(eventActions.addTooltipItem(tooltipItemX));
+
+    const yTooltipItems = ys.map((y, index) => ({
+        datum,
+        name: y,
+        value: datum[y],
+        icon: "square" as const,
+        fill: colors[index],
+    }));
+
+    yTooltipItems.forEach((y) => {
+        dispatch(eventActions.addTooltipItem(y));
+    });
+
+    dispatch(eventActions.setPositionEvent({ x: event.offsetX, y: event.offsetY }));
+
+    return () => {
+        dispatch(eventActions.setTooltipBorderColor(undefined));
+        dispatch(eventActions.removeTooltipItem(tooltipItemX));
+
+        yTooltipItems.forEach((y) => {
+            dispatch(eventActions.removeTooltipItem(y));
+        });
+    };
+}
+
+/**
  * Handles the user interacting with a DataPoint on the Column chart and the need to display a tooltip
  * @param  dispatch        The redux store dispatch function
  * @param  x               The key for the x value
  * @return                 A function to set the tooltip datum
  */
-const useTooltip = ({ x }: Omit<column.IColumnTooltipProps, "dispatch" | "datum" | "colors" | "event" | "ys">) => {
+const useTooltip = ({ x }: Omit<IColumnTooltipProps, "dispatch" | "datum" | "colors" | "event" | "ys">) => {
     const { dispatch } = useStore();
     const [datum, setDatum] = useState(null);
     const [colors, setColors] = useState(null);
@@ -25,7 +96,7 @@ const useTooltip = ({ x }: Omit<column.IColumnTooltipProps, "dispatch" | "datum"
     const [positionEvent, setPositionEvent] = useState(null);
 
     useEffect(() => {
-        return column.tooltip({ dispatch, datum, colors, event: positionEvent, x, ys });
+        return tooltip({ dispatch, datum, colors, event: positionEvent, x, ys });
     }, [dispatch, colors, x, ys, positionEvent]);
 
     /**
